Update patch set 2
Patch Set 2:

> I just stashed, I know all QCOM target use that flag. What about
 > changing it to HAS_S3D_SUPPORT.

They are not quite equivalent. As per the logic we aim for in PA, QTI_BSP things should be runnable on all devices. S3D is a newly introduced feature which works on msm8996(Snapdragon 820) only. 
It is possible that our S3D ifdef is not "perfect", so that can be improved.

Patch-set: 2