Surfaceflinger: adjust content detection fps selection
Select the FPS with minimal error in the first pass and then adjust
based on ratio/margin.

The CL changes content driven FPS selection logic,
before the change, the code switch to 90hz configs
when mContentRefreshRate > 64 and with the CL it will
switch to 90hz when mContentRefreshRate > 75. The second
pass to select 90hz when mContentRefreshRate is 45fps
is kept as is. It also simplified and removed 5% margin
in the first pass.


Test: boot and take SF trace
Bug: 136472613
Change-Id: I4445386a301b7fcdd22d71bd49f45540cc414174