SF: Scheduler: mark keyboard layers as ExplicitIdle
The device will enter the idle state now whenever the keyboard layer is
active.

This slightly differs from forcing this from frameworks/base as the layer
can actively be removed while still being displayed. So, the highest
refresh rate can be still used while the keyboard is still in display.

Change-Id: I25ab3b511f5a8a29177bb8a3c8455171ab85dd8e
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>