SF: Scheduler: enter idle when keyboard is active
The use of the keyboard means the display movement is limited and we can
opportunistically enter idle refresh rate then too for a measurable power
efficiency gain.

The logic is complicated here as it has to overrule touch booster and
keep the refresh rate as low as possible while still being able to boost
when other things happen.

First, the idle refresh rate is entered when 500ms has passed since the
keyboard detection. This is to avoid keyboard animations slowing down.

Also, when keyboard dismissal is detected, it blocks re-entering the
keyboard's idle refresh rate state for 2s because the layer removal is
deferred in Android. Without the blocking logic, the idle refresh rate
will be immediately re-entered due to this.

This entire logic is a little bit conservative, as in, in order to
quickly reach the higher refresh rate states, it'll allow other layers
to disrupt the idle state. This is done by storing the layer state
(savedState in the code) and comparing it later. NoVote, Min and
HeuristicUnresolved (which was previously Max) are considered the same
during this process to only switch when it's quite certain a given layer
wants to use higher refresh rate.

With that said, this design makes a trade-off. Even if you use the
keyboard exclusively without touching anything else, the device will
occasionally enter higher refresh rate states as other layer's states
change. Also, due to how heuristic fps measurement is done with LayerInfo,
the device won't resume to higher refresh rate immediately when e.g., you
scroll without dismissing the keyboard. This is because both layers
(keyboard and the background app) are visible and active, and LayerInfo
must re-calculate its fps to enter higher refresh rate. Note that
creation/deletion of layers will immediately switch to higher refresh
rate, e.g., going to home screen, dismissing the keyboard, etc.

All in all, this design makes a reasonable and practical trade-off and
results in a huge, measurable power efficiency gain.

Change-Id: I806b6ab9ef235d14eb5a9c217b7f30511f3ec8c1
Signed-off-by: Juhyung Park <qkrwngud825@gmail.com>