sf: Perform beginDraw before composition strategy
BeginDraw is executed for draw cycles that are in disabled
output state. This leads to SF queueing setClientTarget_3_1 calls
to HAL unnecessarily. The output state is disabled when power mode
is off.

This change moves beginDraw before the composition strategy selection
to ensure that it only gets called for valid draw cycles.

Change-Id: I437ee50a024c808f49116303c16aa4852daf0077
CRs-Fixed: 3203608