Move absolute axis validation from EventHub to InputDevice
We recently added a crash when a we receive a misconfigured event froma
device. However, it seems like our critera for determining if an
absolute axis is correctly configured has issues.

In this case, we ignore an axis if its min and max value are equal.
Since the min and max are inclusive, it's perfectly valid for them to be
equal. As long as evdev is reporting valid events, EventHub should
continue to track and report them.

Deeming axes with min and max values that are equal is an artificial
constraint we have placed in InputReader historically. To maintain that
same constraint as before, we move the validation to InputDevice.

Bug: 293156873
Test: Presubmit
(cherry picked from https://googleplex-android-review.googlesource.com/q/commit:1030142fbc887c564bb6165946d06ccf43dd9591)
Merged-In: Id7c841f2d06668536a1b61c0e9e5250c15aad550
Change-Id: Id7c841f2d06668536a1b61c0e9e5250c15aad550