[PATCH 1/2] Inline code to turn an SkRuntimeBuilder -> SkImage
Backporting in support of a follow-on commit, which can be submitted to
multiple branches.

Original description:
========================================================================
This appears to be the only use of SkRuntimeEffect::makeImage

This also removes a very small amount of code size and very slightly
reduces parse time by removing whitespace from the shader code.
========================================================================

Fix conflicts. SkSurfaces does not exist in this branch, so use
SkRuntimeEffectBuilder as a guide.

Bug: 290294055
Test: manual (pull down notification shade)
// Change-Id: I51b1b11f1cb49717ab5c069f9f5cb4e66c0b93ff
// Merged-In: I51b1b11f1cb49717ab5c069f9f5cb4e66c0b93ff

Change-Id: Ibddef87968710f6eb63e799dab65bff1779cb301
CRs-Fixed: 3558355