Enlarge the buffer for otapreopt commands.
If a command exceeds the size of the buffer, the return value of
`std::cin.getline` implicitly casts to false, halting the execution.
Logs in a bug report shows that some commands can be longer than 1000
chars because there can be a lot of split APKs in the CLC.

This change enlarges the buffer to work around this issue. Note that
it's not a proper fix to the issue. Instead, it's a quick change to help
us verify the theory in b/307903002.

Bug: 307903002
Change-Id: I27962ce25182c27f9bbbf4016d76aa3f9997eddc
Test: Presubmit