gui: add retry mechanism for GPP session initialization
Some games create the main SurfaceView before the launcher SurfaceView is destroyed. As a result, creating a new GPP session for the main SurfaceView is rejected because the previous session is still running and multiple sessions are not supported.

To address this issue, SurfaceExtensionGPP retries enabling GPP several times. This allows it to successfully create a GPP session for the main SurfaceView after the previous GPP session for the launcher SurfaceView has been destroyed.

CRs-Fixed: 4407450

Change-Id: Ia4657bf697bed532b6c01c7fc7bd1392ffaa8b1e